refactor: muda arquitetura para ficar mais próxima do clean arch
