+++ src/domain/MessageHandler.ts
@@ -1,5 +1,5 @@
 import TelegramBot from 'node-telegram-bot-api';
 
-export interface TelegramHandler {
+export default interface MessageHandler {
     handle(message: TelegramBot.Message): Promise<string>;
 }
