+++ src/infra/adapters/BibleMessageToFullVerseWithReferenceAdapter.ts
